feat: agregar validación para existencia de libro antes de actualizar o eliminar, y filtrar libros eliminados en las consultas

+++ back/src/book-store/aplication/use-case/book-store.usecase.service.ts
@@ -1,4 +1,4 @@
-import { Inject, Injectable } from '@nestjs/common';
+import { Inject, Injectable, NotFoundException } from '@nestjs/common';
 import { CreateBookStoreDto } from '../dto/create-book-store.dto';
 import { UpdateBookStoreDto } from '../dto/update-book-store.dto';
 import { IBookUseCase } from '../interfaces/book-use-case.interface';
@@ -32,10 +32,21 @@ export class BookStoreusecaseService implements IBookUseCase {
   findOne(id: string): Promise<BookModel | null> {
     return this.bookRepository.findOne(id);
   }
-  update(id: string, dto: UpdateBookStoreDto): Promise<{ message: string; }> {
+  async update(id: string, dto: UpdateBookStoreDto): Promise<{ message: string; }> {
+    //validar que el libro exista antes de actualizarlo
+    const book = await this.bookRepository.findOne(id);
+    if (!book) {
+      throw new NotFoundException(`Character with id ${id} not found`);
+    }
+    await this.cacheService.deleteByPattern(`${BOOK_CACHE_KEYS.PATTERN_ALL}`);
     return this.bookRepository.update(id, dto);
   }
-  remove(id: string): Promise<{ message: string; }> {
+  async remove(id: string): Promise<{ message: string; }> {
+    const book = await this.bookRepository.findOne(id);
+    if (!book) {
+      throw new NotFoundException(`Character with id ${id} not found`);
+    }
+    await this.cacheService.deleteByPattern(`${BOOK_CACHE_KEYS.PATTERN_ALL}`);
     return this.bookRepository.remove(id);
   }
